+++ CHANGELOG.md
@@ -3,4 +3,9 @@
 All changes and updates, that are relevant for a user will be documented here
 
 ## Next Release
+
+### Error Handling
+
+- Add Global Exception Dialog for unrecoverable errors
+
 <!-- insertion marker -->